+++ src/components/form/form-preview.vue
@@ -10,13 +10,17 @@
 					Your name
 				</form-input>
 
-				<radio-group v-bind="{ options: options }" v-model="favouriteIceCream">
+				<radio-group v-bind="{ options: selectOptions }" v-model="favouriteIceCream">
 					Favourite ice-cream
 
 					<template #introduction>
 						You will be judged based on your answer.
 					</template>
 				</radio-group>
+
+				<button-group v-bind="{ options: buttonGroupOptions }" v-model="theme">
+					Theme
+				</button-group>
 
 				<form-actions class="mx-[calc(-1.5rem+4px)] mb-[calc(-1.5rem+4px)] rounded-[3px] bg-grey-50 p-[calc(1.5rem-4px)]">
 					<ui-button class="button--primary">
@@ -25,7 +29,7 @@
 				</form-actions>
 			</form-layout>
 
-			<pre class="grow rounded-md border border-grey-200 p-6">{{ { name, favouriteIceCream } }}</pre>
+			<pre class="grow rounded-md border border-grey-200 p-6">{{ { name, favouriteIceCream, theme } }}</pre>
 		</div>
 	</preview-wrapper>
 </template>
@@ -35,5 +39,7 @@ import { ref } from "vue";
 
 const name = ref("");
 const favouriteIceCream = ref("");
-const options = [{ label: "Chocolate", value: "chocolate" }, { label: "Vanilla", value: "vanilla" }, { label: "Strawberry", value: "strawberry" }];
+const theme = ref("");
+const selectOptions = ["Chocolate", "Vanilla", "Strawberry"];
+const buttonGroupOptions = ["System", "Light", "Dark"];
 </script>