+++ src/previews.js
@@ -2,6 +2,7 @@ import PreviewSection from "@/preview/preview-section.vue";
 import PreviewWrapper from "@/preview/preview-wrapper.vue";
 import FormLayout from "@/components/form/form-layout/form-layout.vue";
 
+import ButtonGroupPreview from "@/components/form/button-group/button-group-preview.vue";
 import FormInputPreview from "@/components/form/form-input/form-input-preview.vue";
 import FormPreview from "@/components/form/form-preview.vue";
 import LinkTagPreview from "@/components/interaction/link-tag/link-tag-preview.vue";
@@ -12,13 +13,14 @@ import UiButtonPreview from "@/components/interaction/ui-button/ui-button-previe
 // Prepare previews for use with <component :is> to simplify preview management.
 const previewLibrary = {
 	install(app) {
-		app.component("PreviewSection", PreviewSection);
-		app.component("PreviewWrapper", PreviewWrapper);
-		app.component("FormLayout", FormLayout);
+		app.component("ButtonGroupPreview", ButtonGroupPreview);
 		app.component("FormInputPreview", FormInputPreview);
+		app.component("FormLayout", FormLayout);
 		app.component("FormPreview", FormPreview);
 		app.component("LinkTagPreview", LinkTagPreview);
 		app.component("PillBadgePreview", PillBadgePreview);
+		app.component("PreviewSection", PreviewSection);
+		app.component("PreviewWrapper", PreviewWrapper);
 		app.component("RadioGroupPreview", RadioGroupPreview);
 		app.component("UiButtonPreview", UiButtonPreview);
 	},