+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `link-tag` - A link tag for internal or external links.
 - `form-input`- A text input with extra options.
 - `radio-group`- A group of radio buttons generated from a list of options.
+- `button-group`- A group of buttons generated from a list of options. An extension of `radio-group` with different styling.
 - `conditional-wrapper`- A convenience component to conditionally wrap (or not) content with an element.
 - `form-layout`- A convenience component to help with basic arrangement of form fields.
 - `form-actions`- A convenience component to help with basic arrangement of form actions.